[WFLY-13132] Adjust ejb3 model version from 7.0 to 8.0
