+++ ejb3/src/main/java/org/jboss/as/ejb3/subsystem/EJB3SubsystemNamespace.java
@@ -45,7 +45,8 @@ public enum EJB3SubsystemNamespace {
     EJB3_4_0("urn:jboss:domain:ejb3:4.0"),
     EJB3_5_0("urn:jboss:domain:ejb3:5.0"),
     EJB3_6_0("urn:jboss:domain:ejb3:6.0"),
-    EJB3_7_0("urn:jboss:domain:ejb3:7.0");
+    EJB3_7_0("urn:jboss:domain:ejb3:7.0"),
+    EJB3_8_0("urn:jboss:domain:ejb3:8.0");
 
 
     private final String name;
