+++ ejb3/src/main/resources/schema/wildfly-ejb3_8_0.xsd
@@ -2,7 +2,7 @@
 
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2019, Red Hat, Inc., and individual contributors
+  ~ Copyright 2020, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -23,12 +23,12 @@
   -->
 
 <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
-           targetNamespace="urn:jboss:domain:ejb3:7.0"
-           xmlns="urn:jboss:domain:ejb3:7.0"
+           targetNamespace="urn:jboss:domain:ejb3:8.0"
+           xmlns="urn:jboss:domain:ejb3:8.0"
            xmlns:threads="urn:jboss:domain:threads:1.1"
            elementFormDefault="qualified"
            attributeFormDefault="unqualified"
-           version="7.0">
+           version="8.0">
 
     <xs:import namespace="urn:jboss:domain:threads:1.1" schemaLocation="jboss-as-threads_1_1.xsd"/>
 
@@ -91,15 +91,18 @@
 
     <xs:complexType name="remoteType">
         <xs:all>
-            <xs:element name="channel-creation-options" type="channel-creation-optionsType" minOccurs="0"
-                        maxOccurs="1"/>
+            <xs:element name="channel-creation-options" type="channel-creation-optionsType" minOccurs="0" maxOccurs="1"/>
             <xs:element name="profiles" type="profilesType" minOccurs="0" maxOccurs="1"/>
         </xs:all>
-        <xs:attribute name="connector-ref" type="xs:string" use="required"/>
+        <xs:attribute name="connectors" type="connectorsType" use="required"/>
         <xs:attribute name="thread-pool-name" type="xs:token" use="required"/>
         <xs:attribute name="cluster" type="xs:string" use="optional"/>
         <xs:attribute name="execute-in-worker" type="xs:boolean" use="optional" default="true"/>
     </xs:complexType>
+
+    <xs:simpleType name="connectorsType">
+        <xs:list itemType="xs:string"/>
+    </xs:simpleType>
 
     <xs:complexType name="profilesType">
         <xs:choice minOccurs="0" maxOccurs="unbounded">
@@ -130,8 +133,7 @@
 
     <xs:complexType name="remoting-ejb-receiverType">
         <xs:all>
-            <xs:element name="channel-creation-options" type="channel-creation-optionsType" minOccurs="0"
-                        maxOccurs="1"/>
+            <xs:element name="channel-creation-options" type="channel-creation-optionsType" minOccurs="0" maxOccurs="1"/>
         </xs:all>
         <xs:attribute name="name" type="xs:string" use="required"/>
         <xs:attribute name="outbound-connection-ref" type="xs:string" use="required"/>
