+++ ejb3/src/test/resources/org/jboss/as/ejb3/subsystem/subsystem.xml
@@ -1,4 +1,4 @@
-<subsystem xmlns="urn:jboss:domain:ejb3:7.0">
+<subsystem xmlns="urn:jboss:domain:ejb3:8.0">
     <session-bean>
         <stateless>
             <bean-instance-pool-ref pool-name="slsb-strict-max-pool"/>
